requirements.txt
# twitter-control
python-twitter == 3.1
requests == 2.10.0
pyyaml == 3.11
